The possibility of comments before a scalar has been added.
When building a scalar via a ScalarBuilder, for each type of scalar there is a method to build which receives the boolean inlineComment which determines whether the comment is written inline after the value or before the scalar.
If no inlineComment value is explicitely defined, the method building the scalar checks whether the comment contains a line break. If that's the case, a multiline comment before the scalar is used, if not, an inline comment is used.
If it is explicitely defined that an inline comment should be used which contains line breaks, those are transformed to spaces. Therefore, the previous bug that parts of a comment with line breaks are printed not as comments but as values after the scalar is resolved.
The printer uses this information which is now saved in the comment to print the nodes. Therefore the comments of scalars are printed at the set place now.